changed color
The file for some reason had 2 copies of the code in it, I deleted the extra one.  I played around with some color this morning.  It can definitely be changed I was just messing around while I was looking at the footer.  I am not sure why the lines are at the bottom of the pictures.  I looked at the code and it all looks fine.  Maybe it's just something that is in the image?